TIP-464: add subdirectories Csv, Xlsx and Yaml for Writers
