+++ CHANGELOG-1.6.md
@@ -69,12 +69,9 @@
 - Remove properties editTemplate, showTemplate from `src\Akeneo\Component\Batch\Job\Job`.
 - Remove methods setShowTemplate, setEditTemplate from `src\Akeneo\Component\Batch\Job\Job`.
 - Change constructor of `Pim\Bundle\ImportExportBundle\Controller\JobProfileController`. Add `Akeneo\Bundle\BatchBundle\Connector\JobTemplateProviderInterface`
-- Change constructor of `Pim\Component\Connector\Writer\File\CsvWriter` . Add parameter `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
-- Change constructor of `Pim\Component\Connector\Writer\File\CsvProductWriter` . Add parameter `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
-- Change constructor of `Pim\Component\Connector\Writer\File\CsvVariantGroupWriter` . Add parameter `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
-- Change constructor of `Pim\Component\Connector\Writer\File\XlsxSimpleWriter` . Add `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
-- Change constructor of `Pim\Component\Connector\Writer\File\XlsxProductWriter` . Add `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
-- Change constructor of `Pim\Component\Connector\Writer\File\XlsxVariantGroupWriter` . Add `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
+- Change constructor of `Pim\Component\Connector\Writer\File\Csv\Writer` . Add parameter `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
+- Change constructor of `Pim\Component\Connector\Writer\File\Csv\ProductWriter` . Add parameter `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
+- Change constructor of `Pim\Component\Connector\Writer\File\Csv\VariantGroupWriter` . Add parameter `Pim\Component\Connector\Writer\File\ColumnSorterInterface`
 - Remove method `setAvailableLocales` in `Pim\Component\Catalog\Model\AttributeInterface` and `Pim\Component\Catalog\Model\AbstractAttribute`
 - `Pim\Component\Connector\Writer\File\FlatItemBuffer` implements `\Countable`
 - `Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter\DateFilter` does not implement `Pim\Component\Catalog\Query\Filter\FieldFilterInterface`
