+++ src/Pim/Bundle/BaseConnectorBundle/spec/Archiver/FileWriterArchiverSpec.php
@@ -11,9 +11,8 @@ use Akeneo\Component\Batch\Step\AbstractStep;
 use Akeneo\Component\Batch\Step\ItemStep;
 use League\Flysystem\Filesystem;
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Connector\Writer\File\CsvWriter;
+use Pim\Component\Connector\Writer\File\Csv\Writer;
 use Prophecy\Argument;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 class FileWriterArchiverSpec extends ObjectBehavior
 {
@@ -32,7 +31,7 @@ class FileWriterArchiverSpec extends ObjectBehavior
     function it_creates_a_file_when_writer_is_valid(
         $filesystem,
         $jobRegistry,
-        CsvWriter $writer,
+        Writer $writer,
         JobExecution $jobExecution,
         JobInstance $jobInstance,
         Job $job,
@@ -118,7 +117,7 @@ class FileWriterArchiverSpec extends ObjectBehavior
 
     function it_supports_a_compatible_job(
         $jobRegistry,
-        CsvWriter $writer,
+        Writer $writer,
         JobExecution $jobExecution,
         JobInstance $jobInstance,
         Job $job,
