+++ src/Pim/Component/Connector/Writer/File/Csv/VariantGroupWriter.php
@@ -1,34 +1,39 @@
 <?php
 
-namespace Pim\Component\Connector\Writer\File;
+namespace Pim\Component\Connector\Writer\File\Csv;
+
+use Pim\Component\Connector\Writer\File\BulkFileExporter;
+use Pim\Component\Connector\Writer\File\ColumnSorterInterface;
+use Pim\Component\Connector\Writer\File\FilePathResolverInterface;
+use Pim\Component\Connector\Writer\File\FlatItemBuffer;
 
 /**
- * Write product data into a csv file on the local filesystem
+ * CSV variant group writer
  *
- * @author    Yohan Blain <yohan.blain@akeneo.com>
- * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @author    Samir Boulil <samir.boulil@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class CsvProductWriter extends CsvWriter
+class VariantGroupWriter extends Writer
 {
     /** @var BulkFileExporter */
-    protected $mediaCopier;
+    protected $fileExporter;
 
     /**
      * @param FilePathResolverInterface $filePathResolver
      * @param FlatItemBuffer            $flatRowBuffer
-     * @param BulkFileExporter          $mediaCopier
+     * @param BulkFileExporter          $fileExporter
      * @param ColumnSorterInterface     $columnSorter
      */
     public function __construct(
         FilePathResolverInterface $filePathResolver,
         FlatItemBuffer $flatRowBuffer,
-        BulkFileExporter $mediaCopier,
+        BulkFileExporter $fileExporter,
         ColumnSorterInterface $columnSorter
     ) {
         parent::__construct($filePathResolver, $flatRowBuffer, $columnSorter);
 
-        $this->mediaCopier = $mediaCopier;
+        $this->fileExporter = $fileExporter;
     }
 
     /**
@@ -36,32 +41,25 @@ class CsvProductWriter extends CsvWriter
      */
     public function write(array $items)
     {
-        $products = $media = [];
-        foreach ($items as $item) {
-            $products[] = $item['product'];
-            $media[] = $item['media'];
-        }
-
-        parent::write($products);
-
-        $parameters = $this->stepExecution->getJobParameters();
-
-        if ($parameters->has('with_media') && !$parameters->get('with_media')) {
-            return;
-        }
-
         $exportDirectory = dirname($this->getPath());
         if (!is_dir($exportDirectory)) {
             $this->localFs->mkdir($exportDirectory);
         }
 
-        $this->mediaCopier->exportAll($media, $exportDirectory);
+        $variantGroups = $media = [];
+        foreach ($items as $item) {
+            $variantGroups[] = $item['variant_group'];
+            $media[]         = $item['media'];
+        }
 
-        foreach ($this->mediaCopier->getCopiedMedia() as $copy) {
+        parent::write($variantGroups);
+        $this->fileExporter->exportAll($media, $exportDirectory);
+
+        foreach ($this->fileExporter->getCopiedMedia() as $copy) {
             $this->writtenFiles[$copy['copyPath']] = $copy['originalMedium']['exportPath'];
         }
 
-        foreach ($this->mediaCopier->getErrors() as $error) {
+        foreach ($this->fileExporter->getErrors() as $error) {
             $this->stepExecution->addWarning(
                 $error['message'],
                 [],
