--- src/Pim/Component/Connector/Writer/File/CsvVariantGroupWriter.php
@@ -1,65 +1,0 @@
-<?php
-
-namespace Pim\Component\Connector\Writer\File;
-
-/**
- * CSV variant group writer
- *
- * @author    Samir Boulil <samir.boulil@akeneo.com>
- * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class CsvVariantGroupWriter extends CsvWriter
-{
-    /** @var BulkFileExporter */
-    protected $fileExporter;
-
-    /**
-     * @param FilePathResolverInterface $filePathResolver
-     * @param FlatItemBuffer            $flatRowBuffer
-     * @param BulkFileExporter          $fileExporter
-     * @param ColumnSorterInterface     $columnSorter
-     */
-    public function __construct(
-        FilePathResolverInterface $filePathResolver,
-        FlatItemBuffer $flatRowBuffer,
-        BulkFileExporter $fileExporter,
-        ColumnSorterInterface $columnSorter
-    ) {
-        parent::__construct($filePathResolver, $flatRowBuffer, $columnSorter);
-
-        $this->fileExporter = $fileExporter;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function write(array $items)
-    {
-        $exportDirectory = dirname($this->getPath());
-        if (!is_dir($exportDirectory)) {
-            $this->localFs->mkdir($exportDirectory);
-        }
-
-        $variantGroups = $media = [];
-        foreach ($items as $item) {
-            $variantGroups[] = $item['variant_group'];
-            $media[]         = $item['media'];
-        }
-
-        parent::write($variantGroups);
-        $this->fileExporter->exportAll($media, $exportDirectory);
-
-        foreach ($this->fileExporter->getCopiedMedia() as $copy) {
-            $this->writtenFiles[$copy['copyPath']] = $copy['originalMedium']['exportPath'];
-        }
-
-        foreach ($this->fileExporter->getErrors() as $error) {
-            $this->stepExecution->addWarning(
-                $error['message'],
-                [],
-                $error['medium']
-            );
-        }
-    }
-}
