+++ src/Pim/Component/Connector/Writer/File/Yaml/Writer.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace Pim\Component\Connector\Writer\File;
+namespace Pim\Component\Connector\Writer\File\Yaml;
 
 use Akeneo\Component\Batch\Job\RuntimeErrorException;
+use Pim\Component\Connector\Writer\File\AbstractFileWriter;
+use Pim\Component\Connector\Writer\File\FilePathResolverInterface;
 use Symfony\Component\Yaml\Yaml;
 
 /**
@@ -12,7 +14,7 @@ use Symfony\Component\Yaml\Yaml;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class YamlWriter extends AbstractFileWriter
+class Writer extends AbstractFileWriter
 {
     const INLINE_ARRAY_LEVEL = 8;
