+++ src/Pim/Component/Connector/spec/Writer/File/Xlsx/SimpleWriterSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Writer\File;
+namespace spec\Pim\Component\Connector\Writer\File\Xlsx;
 
 use Akeneo\Component\Batch\Job\JobParameters;
 use Akeneo\Component\Batch\Model\StepExecution;
@@ -10,7 +10,7 @@ use Pim\Component\Connector\Writer\File\FlatItemBufferFlusher;
 use Pim\Component\Connector\Writer\File\FlatItemBuffer;
 use Prophecy\Argument;
 
-class XlsxSimpleWriterSpec extends ObjectBehavior
+class SimpleWriterSpec extends ObjectBehavior
 {
     function let(
         FilePathResolverInterface $filePathResolver,
@@ -26,7 +26,7 @@ class XlsxSimpleWriterSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Pim\Component\Connector\Writer\File\XlsxSimpleWriter');
+        $this->shouldHaveType('Pim\Component\Connector\Writer\File\Xlsx\SimpleWriter');
     }
 
     function it_is_step_execution_aware()
